weed
bud-core (B.U.D. 1.0/2.0/3.0, 103 modul) CI'da hic derlenmiyordu: ci.yml'e uc yapilandirma (default/bud3/no-default) clippy -D warnings + test adimi baglandi; clippy 91 ihlal -> 0; exepdf.rs if'in iki dali ayni idi (icerik-duyarli bolumleme iddiasi sahte, olcum hesaplanip atiliyordu) sabit oranli hale getirildi; logfield.rs vec![Vec::with_capacity(n); 7] kapasiteyi 7 sutunun 6'sinda kaybediyordu; lib.rs cift cfg yuzunden bud_format_qrvideo ve bud_format_sertlesme 3.0 kapisinin disinda kalmisti (bud2 derlemesine 3.0 kodu siziyordu); pgm.rs never_loop tek gecisli dal if'e cevrildi; securededup.rs explicit_counter_loop; revolutionary.rs inherent to_string -> Display; price.rs media_device_only_holds assert!(0.0 <= 0.016) iki sabiti karsilastiriyordu, model uzerinden gercek teste cevrildi + kanarya; test_bud2.rs assert!(true) yer tutucusu 6 kayipsizlik degismezi testiyle degistirildi (mutasyon kanaryasi ile dogrulandi: tek bit bozulmada 4 test kirmizi); bud/Cargo.toml [lints.rust]/[lints.clippy] kapisi + bud1/bud2 feature'larinin cfg kapisi OLMADIGI belgelendi; 490/490 test gecer

+++ bud/src/bud_format_por.rs
@@ -57,7 +57,7 @@ impl PorKey {
         }
         let mut nonce = [0u8; 32];
         nonce[0..8].copy_from_slice(&seed.to_le_bytes());
-        nonce[8..16].copy_from_slice(&(block_count as u64).to_le_bytes());
+        nonce[8..16].copy_from_slice(&block_count.to_le_bytes());
         PorChallenge { indices, nonce }
     }
